Add Not operator for pre-requisites
This allows tasks (features) that are mutually exclusive. For example, when an app has basic UI, it cannot also have “programmer art” UI, and vice versa. This should make the decision tree cleaner, while allowing widely divergent apps to be built.

This also fixes a bug with AllOf() and AnyOf().