Set sustained performance mode
This makes the app run fast even when not being touched. We should have this for the demo at I/O. On the other hand, the released app should _not_ have this.

https://developer.android.com/about/versions/nougat/android-7.0.html#sustained_performance_api

Alternatively, we can have this in settings. But that requires building a plugin / calling to Kotlin, which is work we haven’t bargained for.

Note that this does not fully prevent Android from throttling CPU according to load. In other words, this is not sufficient for performance testing. This is only to make sure a demo device in demo mode runs at optimal speed even when it hasn’t been touched by any user for a long time.